requirements.txt
sentry-sdk==1.5.4
mysqlclient==2.1.0
djangorestframework==3.13.1
gunicorn==20.1.0
python-keystoneclient==4.3.0
python-novaclient==17.6.0
python-glanceclient==3.5.0
django-tables2==2.4.1
python-dateutil==2.8.2
keystoneauth1==4.4.0
